Spells are now read/written from/to the Google Sheets. Added Index to Spell and SpellDto classes. Added secondary hue, sat, and lightness (for Fireball effect) to the AddSpellHit expression function (it takes up to 11 parameters now). Changed ApplyCommand to now collect a number of any size. Added Digit and All Clear ("ClearNumbers") Stream Deck commands. The Digit Stream Deck command can replace "{count}" appearing in other Stream Deck commands, like InstantRoll. Added WildShape property to Character.cs. Updated replaced Wild Shape properties to pull from the Wild Shape. Added FloatPlayerText SignalR command to float text from the bottom up to the top of the screen at the specified player's X position. Added ability to invoke evaluated expressions before executing the Stream Deck Command - just add the expression after a " | " at the end of the Stream Deck command. Saving Features back to the Google Sheets document. Added the starts of a Monster picker. GoogleHelper can now match columns ignoring case. Added SetNextAnswer function to the Expression engine, allowing Stream Deck commands to answer future UI questions automatically. Added WildShapeTests.

+++ DHDM/IDungeonMasterApp.cs
@@ -7,7 +7,7 @@ namespace DHDM
 {
 	public interface IDungeonMasterApp
 	{
-		void Apply(string command, int value, List<int> playerIds);
+		void Apply(string command, decimal value, List<int> playerIds);
 		string GetPlayFirstNameFromId(int playerId);
 		void SetBoolProperty(int playerId, string propertyName, bool value);
 		bool GetBoolProperty(int playerId, string propertyName);
